DESCRIPTION For safe and secure lending experience, it's important to analyze the past data.
DESCRIPTION
For safe and secure lending experience, it's important to analyze the past data. In this project, you have to build a deep learning model to predict the chance of default for future loans using historical data. As you will see, this dataset is highly imbalanced and includes a lot of features that make this problem more challenging.

Objective: Create a model that predicts whether or not an applicant will be able to repay a loan using historical data.
Domain: Finance
Analysis to be done: Perform data pre-processing and build a deep learning prediction model.
Steps to be done: 
⦁    Load the dataset that is given to you
⦁    Check for null values in the dataset
⦁    Print percentage of default to payer of the dataset for the TARGET column
⦁    Balance the dataset if the data is imbalanced
⦁    Plot the balanced data or imbalanced data
⦁    Encode the columns that is required for the model
⦁    Calculate Sensitivity as a metrics
⦁    Calculate area under receiver operating characteristics curve